refactor: Move group info handling into own module.
