+++ openmls/src/framing/message_out.rs
@@ -11,7 +11,11 @@ use tls_codec::Serialize;
 
 use super::*;
 
-use crate::{key_packages::KeyPackage, versions::ProtocolVersion};
+use crate::{
+    key_packages::KeyPackage,
+    messages::group_info::{GroupInfo, VerifiableGroupInfo},
+    versions::ProtocolVersion,
+};
 
 /// An [`MlsMessageOut`] is typically returned from an [`MlsGroup`] function and
 /// meant to be serialized and sent to the DS.
