+++ openmls/src/group/core_group/mod.rs
@@ -45,8 +45,11 @@ use crate::{
     framing::{mls_auth_content::AuthenticatedContent, *},
     group::*,
     key_packages::*,
-    messages::VerifiableGroupInfo,
-    messages::{proposals::*, *},
+    messages::{
+        group_info::{GroupInfo, GroupInfoTBS, VerifiableGroupInfo},
+        proposals::*,
+        *,
+    },
     schedule::{message_secrets::*, psk::*, *},
     tree::{secret_tree::SecretTreeError, sender_ratchet::SenderRatchetConfiguration},
     treesync::{
